add parameter types

+++ src/Controller/InfoSheetsController.php
@@ -88,7 +88,7 @@ class InfoSheetsController extends AppController
             'workshopUid' => $workshopUid
         ];
         $filename = 'Laufzettel-Download-' . StringComponent::slugifyAndKeepCase($workshop->name);
-        if (in_array($year, Configure::read('AppConfig.timeHelper')->getAllYearsUntilThisYear(date('Y'), 2010))) {
+        if (in_array($year, Configure::read('AppConfig.timeHelper')->getAllYearsUntilThisYear((int) date('Y'), 2010))) {
             $query = preg_replace('/WHERE 1/', 'WHERE 1 AND DATE_FORMAT(e.datumstart, \'%Y\') = :year', $query);
             $params['year'] = $year;
             $filename .= '-' . $year;
